Adds a signal's metadata attribute
This adds a signal's metadata attribute which lets you store
any kind and number of metadata attributes to your signal. These
metadata are just reissued at the event emission without alteration.

Special case of "unit" attribute which is kept as a signal attribute
and not as a metadata because it is an intrinsic signal attribute that
all signals should have.

Bug-AGL: SPEC-1419

Change-Id: I531b4012ac9b22f312a4ccebc1a07a8d93129403
Signed-off-by: Romain Forlot <romain.forlot@iot.bzh>