Add steering wheel and non-diagnostic vehicle status signals
Add definitions for the steering wheel button CAN signals to provide
an abstraction for application usage. Also add signals based on the
non-diagnostic vehicle speed, engine speed, and fuel CAN signals to
allow abstracting usage of those by e.g. the dashboard application.

In addition, SourceAPI::addSignal in source.cpp required a tweak to
allow subscribing to signals without specifying an explicit action,
i.e. getting the default on receive behavior as documented. Without
this change, the signals will silently not be subscribed to, and no
events are generated. Since there is no longer an explicit built-in
plugin function that could be called in an action to yield the same
behavior (it was removed in commit 3e45dda and not restored with the
other built-in function by commit bf964e8), I believe this restores
the intended default behavior.

Bug-AGL: SPEC-3024

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: I86e8be66ff27484b34a52145b969661277332a50