[signal]Always call the default callback
Always call the default callback when receiving a signal from a source. This is this callback
that set the timestamp and value attribute of the signal. So we need to always call it else the
attributes are empty and unusable in plugin.

This is a expected behavior so don't let the user in charge of this default behavior.

Change-Id: I6ec23169861e795baa855dd85767638e5d7260ff
Signed-off-by: Romain Forlot <romain.forlot@iot.bzh>