[signal]Improve json_object reference management
We got 3 holders of the value object once recorded in the signal object (which is too much) and
every time we get the value using the get verb the ref count was decreased then it could not be
retrieved no more than 2 times.
This fix the bug and also remove an unecesserary increment.

Change-Id: Ie90a033938108e16ebd3945860370ec7930fcd4d
Signed-off-by: Romain Forlot <romain.forlot@iot.bzh>